Removing strict dependency on the default mod
If one should desire to use this mod for liquid in a
non-default-dependent game, I've made the dependency on default
optional. Note that the only thing this mod does in that case is define
an API, all the rest of its features are moot (since they act on
default-defined liquids and mapgen clay).